Move conversion table spec.
